+++ package.json
@@ -63,6 +63,7 @@
         "graphql-tag": "^2.11.0",
         "graphql-ws": "^5.11.2",
         "js-yaml": "^3.13.1",
+        "jsonwebtoken": "^9.0.2",
         "node-localstorage": "^2.2.1",
         "nodemon-webpack-plugin": "^4.3.2",
         "react-dom": "^18.2.0",
